refactor: migrate link
