+++ packages/components/link/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './link';
+
+export { default as version } from './version';
