+++ packages/components/link/src/link.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../../test-utils';
+import { render } from '../../../../src/test-utils';
 import Link from './link';
 
 const createTestProps = (custom) => ({
