+++ packages/components/link/src/link.story.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { withKnobs, text, boolean } from '@storybook/addon-knobs/react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
+import Readme from '../README.md';
 import Link from './link';
 
 storiesOf('Components|Links', module)
